requirements.txt
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
celery==5.4.0
redis==5.0.8
mysqlclient>=2.0
confluent-kafka==2.5.3